로그인, 로그아웃, 상품목록 -- 전부 미완성

+++ SHOP_LHM/src/main/java/shop/dto/Product.java
@@ -95,7 +95,7 @@ public class Product implements Serializable {
 		return unitsInStock;
 	}
 
-	public void setUnitsInStock(long unitsInStock) {
+	public void setUnitsInStock(Integer unitsInStock) {
 		this.unitsInStock = unitsInStock;
 	}
